RELATED: BB-1159 Refactor of the derived measures title factory
The factory should not be dependent on the measure definition (similarly to the arithmetic measure title factory). This way the whole measure definition does not need to be stored in the decorated bucket item in Analytical Designer. For this purpose the OverTimeComparison interface was moved from gdc-app-components into the gooddata-react-component repo.