RELATED: BB-1203 Migrate date filters merging from web client to backend
The logic of the date filter merging is moved from frontend to the backend as the backend needs to know which filters were global and which local (measure filters) in order to construct correct MAQL queries.

In addition to that, the backend REST API is versioned now. The original API version 1 still expects the date filters to be merged by the frontend (in order for GoodData.UI SDK to be backward compatible for at least a year) while API version 2  and newer expects the filters in the original buckets as they were set by the user.

When the backend returns X-GDC-DEPRECATED response header, the frontend client will output warning into the browser's console. The header is returned in the case when value sent in the X-GDC-VERSION does not identify the latest REST API version deployed on the backend server.