Generate rec index into playlist, use the index in recorded backend impl
Playlist generator now creates recording Master Index in the playlist.ts
This is a map of workspace > exec footprint > recording.
This map is input to recordedBackend(); the SPI method implementations
use this map once execute() is called on prepared execution.. they build
the appropriate result. The execute bombs if recording is not found