Fix pluggable pivot table drilling
As the CorePivotTable is checking the output of the onDrill function
before it decides whether to fire the event, we need to return
the value in teh onDrill in AbstractPluggableVisualization.
Moreover, CorePivotTable onDrill defaults to always true, but now that
we always provide a callback when used from PluggablePivotTable
this default is ignored. Hence we need to default to returning true
and make the drilling explicitly opt-out (which it was implicitly anyway).

JIRA: RAIL-2121