Improve the UMD build logic in bear-client
There are several things going on here:
* LodashModuleReplacementPlugin was fixed so that lodash/fp packages work
* deprecated awesome-typescript-loader was replaced by ts-loader
* dist folder was renamed to umd to reflect its purpose better (it contains an UMD version of the build)
* lib folder was renamed to dist so that its semantics are the same as for the other packages in the monorepo

Webpack was kept as a builder mainly because of the LodashModuleReplacementPlugin.
This plugin makes the UMD bundle half the size and has no counterpart in other bundlers (Rollup, Parcel).
We may revisit this later, but for now we keep webpack as getting rid of it would be lots of work.

JIRA: RAIL-2276