Add support for complex objects to tiger client, use it in insight search
This adds support for passing complex objects as tiger client parameters
and them being properly passed to the backend. This is done by a bit of qs
trickery explained in the template comment.

The new support is used in tiger insight search so that we support LIKE searching.

JIRA: RAIL-2436