Fix cached executions' definitions
Makes sure that cached executions are returned with the definitions used to request them.
This is needed, because definitions with different bucket configurations
can lead to identical executions.
For example:
definition with measures A, B in the "measures" bucket
vs.
definition with measure A in "measures" bucket and measure B in "secondary_measures" bucket.

The cache used to return the cached item with the "old" definition.
This effectively meant that the definition received could be different from
(and potentially incompatible with) the definition used to make the request.
This is a problem because in some contexts the definition is read from the dataView,
and if it is not compatible with the request definition, it can cause logic problems.

This change makes sure that even cache hits are returned
with the incoming definitions which are trivially compatible.

JIRA: RAIL-2382, RAIL-2410