Prevent updates after unmount in InsightView
It can happen so the call to updateVisualization happens after unmount
because it waits on async operations. Thus we need to check
that the target DOM node still exists when calling update.
This is done by keeping a ref to the div that is cleared on unmount automatically.

JIRA: RAIL-2407