Fix date dataset availability in tiger
Reflect the fact the date datasets' availability is dictated
by their attribute availability in tiger. This means that
the relevantRestrictingTypes will now never be empty,
so the branch handling that was also removed.

Finally, one now unnecessary cast to any was removed
and an appropriate Error type was used.

JIRA: RAIL-2509