Remove temporary measure format override
When implementing custom measure format we found out that some measures have incorrect format saved in viz. object. As we started to respect the format saved in viz. object the measures would be formatted wrongly so we introduced this fix for them. These viz. objects have been fixed so that the override is not needed any more