Tiger: Load the date dataset (attribute) in catalog
There is a bug in loading attributes from backend (see MDC-354)
which is possible to workaround by providing bigger size limit.

Overall paging does not work well (yet).

Besides that the Date is not a groupable item in the catalog so it has to be
filtered out in grouping function.

JIRA: RAIL-2822