Prevent DashboardView reloads
This makes sure that DashboardView is not reloaded unnecessarily
when only the isScheduledMailDialogVisible prop changes.

The DashboardRenderer was converted from arrow function to normal function
so that its displayName is set automatically (arrow functions do not do that).
This change was adopted in other existing components to reduce redundancy.

Also, an example of email scheduling for DashboardView was added.

JIRA: RAIL-2884