Change the filters prop semantics in DashboardView
Now if the user specifies the filters prop, it is used to override
the filters in dashboard.

This is needed so that the user can really override the existing filters
when using the onFiltersChange prop (without this change the dashboard
filters would still be used making the override impossible).

To enable the user to use the merging logic if they want,
a convenience function was added and documented.

JIRA: RAIL-2847