Add read only prop to DashboardView and use it in examples
Since the mechanism relies on passing a prop anyway, there
is no need to create a special build of examples (since missing
the prop would disable the readonly mode anyway).
Instead, we just use the appropriate flag in all examples.

Also, a bug in DashboardItemWithKpiAlert was fixed related to read only.
This bug does not manifest in KPI Dashboards because the readonly prop
is reflected in the canSetAlert prop as well.

JIRA: RAIL-2931