Add withMeasures method to IElementsQuery
Add a way to limit the elements by a set of measures.
This limits the elements to only ones for which the set of measures has data.

Also mark restrictiveDefinition and restrictiveDefinitionContent
as deprecated - it is hard to set them correctly
and the withAttributeFilters and withMeasures provide a better alternative.

JIRA: RAIL-2079