Provide info about drill definitions on drill
This allows the user to access the original DrillDefinitions when
an implicit drill is triggered. Since there can be more than one
definition responsible for a given drill event, all relevant are included.

Since there is no DrillDown definition, a special DrillDefinition-like
type was defined and used.

The DrillDefinitions are returned as-is (so no custom URL resolution
is done for example). This might be improved later.

JIRA: RAIL-2890