Prevent pivot table growing in height on its own
There is sometimes exactly 2px discrepancy between the height of
gd-table-dashboard-wrapper (set to 100%)
and gd-table-component that gets its height from its parent's maxHeight.
These 2px create a positive feedback loop which causes the table to
grow in height in 2px increments until it reaches its full height.

To avoid this, the maxHeight is reduced by the 2 pixels which avoids
the discrepancy.

This is as a symptomatic fix as they come, but I was unable to find
the reason why it is always exactly 2 pixels and what causes them.

Also "fixing" this in the PluggablePivotTable prevents impact on any
usage of the PivotTable outside of dashboards environment.

JIRA: RAIL-2993