Prevent pivot table growing in height on its own
There is sometimes exactly 2px discrepancy between the height of
gd-table-dashboard-wrapper (set to 100%)
and gd-table-component that gets its height from its parent's maxHeight.
These 2px create a positive feedback loop which causes the table to
grow in height in 2px increments until it reaches its full height.

To fix this, changes to desiredHeight smaller than 2 pixels
are ignored which breaks the loop.

This is as a symptomatic fix as they come, but I was unable to find
the reason why it is always exactly 2 pixels and what causes them.

This fix is located in CorePivotTable so that all the magic
with desired height is collocated to one place.

JIRA: RAIL-2993