Fix tiger totals conversion
Make sure the totals added to the descriptor are unique.

It can happen that there is more than one total of the same type
for a given attribute, but the server will only return data for
the given total type once and there were mismatches in collection
lengths: for example two rows of totals for three total descriptors
(two of which are the same) which broke the components down the road.
So we make sure to only send unique types forward to avoid this mismatch
(effectively mimicking the backend which seems to be de-duplicating
the totals too).

JIRA: RAIL-3246