Fix normalizingBackend for NoDataErrors
Make sure the dataViews passed in the NoDataErrors are denormalized
to ensure that only denormalized data views come from the executions

To test this, dummyBackend had to be extended to allow different
NoDataErrors to be thrown (with or without dataViews).
Using this, some tests were extended to cover both cases.

JIRA: RAIL-3241