sdk-backend-spi: include date granularities
API for loading of date datasets allows to specify additional date granularities that will be fetched with the ones backend returns by default.

The implementation is done only for Bear backend for now. The Tiger have a different functionality to achieve this (setting of the required returned date granularities via modeler).

JIRA: BB-2865