Prevent use of lodash/get
This adds a dep-cruiser rule that disallows the use of lodash/get.
The reasoning is that this function destroys type information (especially
when using the dot notation) and makes refactoring harder. Moreover,
with ?. and ?? operators available, it makes even less sense to use it.

This migrates alle the packages with two exceptions:

* sdk-ui-ext (it has the get key syntax baked in the supported properties logic)
* api-client-bear (there is some key based magic I was too afraid to touch)

I will revisit these in a separate PR and try to limit the use of lodash/get
there as much as possible.

JIRA: FET-723