Fix pivot tables with multiple sorts
If a pivot table has more than one sort and their order is different
than the order of columns, the mismatch triggers an infinite loop.

This makes sure the order of sorts is the same as the order of columns
(not the order in which the col defs were created).

JIRA: RAIL-3319