Upgrade node-fetch and isomorphic-fetch
Makes sure that both node-fetch and isomorphic-fetch are up-to-date
in our depenendencies. This change is safe:
* breaking changes in node-fetch do not affect us
* the browser fetch used by isomorphic-fetch is the same
  as in the previous version of it (3.6.2)

This will unfortunately not solve the security issue because
fixed-data-table-2 has a transitive dependency (unused)
on node-fetch 1.7.3.

However, this makes the wrong node-fetch version bound to only
fixed-data-table-2 so once it is upgraded, the security issue will be
solved automatically (there is a ticket for the upgrade already).

JIRA: RAIL-3311