Close menu when outside click cancels event bubble
The original solution was easier but failed when outside click stopped click event propagation. Menu was not closed in this case.

The comment in the code described alternative solution that handle it. The new reworked click handling should work now also in this rare case. The original explanation comment was left in place (and slightly altered to match the new situation) as it adds a lot of context and information for future component maintenance.

JIRA: BB-2940