Avoid changing visualization-uri-root div
Changing the div can have unforeseen consequences like the visualization
rendering multiple times. So we avoid it and implement the concern
this was introduced for in the parent component (making sure that
loading component is 100% of the height if shown).

Also, avoid passing irrelevant props to InsightRenderer from InsightView.
This makes the InsightRenderer re-render only if needed.

JIRA: RAIL-3488