Prevent multiple concurrent exports of one insight
This mimics the behavior of gdc-dashboards and improves the UX
by actually disabling the buttons while an export is in progress:
gdc-dashboards leave them enabled and clicking them is a noop
while an export is in progress.

To prevent strange closes of the menu when download of the result
file starts (which involves clicking on a hidden link), the Bubble API
was extended to allow us to configure the ignoreClicks* props of
the underlying Overlay and thus ignore this hidden "click".

JIRA: RAIL-3671