Prevent widget reloads on ignored filters changes
If a filter that is ignored changes, the widget does not need to run
the filters query again, its results will be the same. To prevent these
wasteful reloads (which also make the UX worse by showing unnecessary
loading animations), the set of non-ignored filters is "cached" for each
widget and this value is then used synchronously (so no loading) to
decide if a new run of the full filters query is necessary.

To get the non-ignored filters, we use the filters query and override
the widget level filters with an empty array so that we only get
dashboard-level filters (without the potential insight-level ones).

Also, a digest of the filters is used to only reload the non-ignored
filters query only when the dashboard filters changed in a way that can
affect the non-ignored filters (the filters' display forms/datasets,
not their selected values/date config).

JIRA: RAIL-3682