Fix useDashboardQueryProcessing cancelling
We only want to cancel the inflight queries on component unmount to
prevent the react warning, so there is no need to expose the cancel
callback (which was used incorrectly in useWidgetBrokenAlertsQuery,
because it was called on *every* run of the query, not just on unmount).

Now the cancel is handled internally (as it is an internal implementation
detail of that hook) and correctly (only being run on unmount).

JIRA: RAIL-3680