Add support for customization of the Insight menu
Add support for changing the menu shown for insight widgets by default.

There are two different implementations:
1. LegacyDashboardInsightMenu - this is the "old" export widget data menu
   it allows no external customizations.
2. DefaultDashboardInsightMenu - this is the "new" menu that allows more
   than just the export options. It is also open for custom items.

The logic of which implementation is used is as follows:
* if no user customizations are provided, LegacyDashboardInsightMenu
  is used
* else the DefaultDashboardInsightMenu is used

This is done so that the behavior is consistent with gdc-dashboards.

A new DashboardCustomizationsContext was added to store non-component
customization settings (currently only the insightMenuItems).

Localizations and CSS were sourced from gdc-dashboards.

JIRA: RAIL-3627