Fix filter merging in widget filters query
When insight has a date filter set to something, and is used
on a dashboard (and does not ignore date filter there), then
when dashboard date filter was set to All time, original insight
date filter would be used.

That is wrong, it should be overriden by the dashboard level
All time filter.

JIRA: RAIL-3805