NAS-2440 - Implement native dom api functions on pure input
Implementation of native methods from interface IDomNative on PureInput
Added IDomNativeProps to extends props on PureInput
This can be used in another native like elements

New property "nativeLikeAutofocus" that try to focus element after is visible on page (is not hidden and has some size). This optional must be enabled to start behave this new way.

Also, component PureInput clean timers and intervals to if is destroyed.

JIRA: NAS-2440