Escape HTML tags in sdk-ui-ext translations
Since these strings are used in UiConfig and have to be a string,
the easiest way is to escape the spans and keep them as is.

Ideally, we should use the Rich formatting feature of react-intl,
however getting the results back as string (and not ReactNode)
is quite a hassle not worth doing at this point.

JIRA: FET-529