Implement LRUCache and use it
Since the lru-cache in version 4 (the last compatible with IE11) needs
specific setup when used with Webpack 5 (process polyfills), it makes
GoodData.UI unusable with create-react-app 5 that migrated to Webpack 5
and makes it impossible to alter these settings out of the box.

As a workaround, we wanted to migrate to another LRU implementation,
however none of the reasonably used existing implementations were good
for our particular use cases (IE11 support and maxAge):

* quick-lru: does not support IE11 because it uses generators.
  Older versions do not have maxAge which we need
* tiny-lru: does not work with cache size 1 (I sent a PR to fix that
  but I do not expect that to be merged, the package is quite stale)
* hashlru: does not have maxAge

So I ended up implementing a LRU cache in our code base. The code uses
the hashlru algorithm and is heavily inspired by quick-lru (which we
should migrate to once IE11 is dropped and get rid of the version
added in this commit).

One particular caveat of the hashlru algorithm (that is not that
important in my opinion) is that the maxSize is not precise:
the cache can hold more keys than the maxSize, but not by much (it very
much depends on the particular sequence of inserts and evicts). It will,
however, never be more than twice the maxSize. Since we use the maxSize
as more of a "do not cache millions of items here" kind of mechanism,
I think this caveat is ok for us given the performance gains
and relative code simplicity.

JIRA: RAIL-2928