Make DefaultDashboardInsightWidget more robust
Similar to #2041 there are hard-to-reproduce situations when a code
assumes the insights are loaded even thoug for some reason they are
in fact not ready. This results in hard to reproduce issues.

As a fix, we prevent these invariants from happening and log
the occurences. This also avoids explicit not-null assertion.

Also fix two Sonar-reported code smells in the file.

JIRA: RAIL-3925