Prefix ids with workspace id on tiger
This makes the output code more portable: you can then use the ids as-is
in a child workspace as well as the original workspace.

A good example of hy this is useful is a dashboard plugin created in
a parent workspace that should work in the child as well: in that case
the ids must include the prefix from the parent, otherwise the plugin
will break in the child workspace.

After these changes, the ids are always prefixed which makes them
more usable in child workspaces as well.

JIRA: RAIL-4018