Avoid using storybook.raw()
There are breaking changes in the raw() output so we use the more stable
getStorybook function. This should make us more future-proof.

This still uses the deprecated configure() to force the load, we need
to avoid that as well in order to to able to use Storybook 7.

This is still better than before though: we can upgrade to latest
Storybook resolving quite a few security advisories.

JIRA: FET-922