Add internal barrel to sdk-ui-dashboard
This is the first step towards getting rid of DashboardView:
gdc-dashboards depend on some internals, so we need to
export them from their new locations in sdk-ui-dashboard.

We had to add getBrokenAlertFiltersBasicInfoLegacy because
the new one would be very hard to use in gdc-dashboards as is
and it could potentially break existing stuff there.

JIRA: RAIL-4046