Add support for hour and minute granularities to date filter
Date filter component supports hour and minute granularities now in relative presets and relative floating range.

Order of the granularities of relative date presets and of tabs in relative floating range was reversed from the lowest granularity to the lowest one. Reason for this is to have the newly supported granularities at the bottom as they will be used the least often.

The functionality is mainly meant for Tiger platform that supports time granularities. The functionality will be available on Bear platform as well technically but Bear metadata do not allow saving date filter configs with time related granularities (because its time dimension does not support it), therefore no feature flag check is necessary in the component.

JIRA: TNT-680, TNT-682