Preserve ajaxSettings when using withTelemetry on bear
We want to keep whatever ajaxSettings the user did on the original
backend instance to be carried over to the new one when creating
it using the withTelemetry function.

The other copying function still do not keep this as they may affect
authentication and this may mean the ajaxSettings are no longer
relevant in the context of the new instance.

JIRA: RAIL-4096