Rework how plugins define their dependencies
This prevents issues when using singletons, now we let webpack negotiate
the best version (and it chooses only one if it can so no need for
singleton). Singleton: true made the first loaded version get stuck in
so that changing to a dashboard that needs a newer version did not work
as the older one was stuck there and was not overwritten.

This solution does not work with prerelease versions because semver
does not have query meaning:
"Give me a range of versions including prereleases"
so we allow the user to opt out of the version checking if they want to
use prerelease versions (or more specificaly combine plugins with
different prerelease versions on the same dashboard).

JIRA: RAIL-4117