Use defaultView property to pick the default label
This makes use of the new defaultView relationship to pick the default
label for a given attribute. If no defaultView is specified, we fall
back to the primary flag.

Also fix the dataset include, the datasetS version is obsolete and
does not work in conjunction with the defaultView include.

JIRA: RAIL-4178