Reflect defaultView in metadata loading
We must change the way a single display form is loaded because
loading the label directly does not contain the information whether
it is its attribute's defaultView.

Instead, we use RSQL to load the attribute and then find the requested
label in its labels. We also remove several obsolete convertors.

JIRA: RAIL-4210