Avoid TS errors in unused files in playground
This makes the output only warn about errors in the files actually used
in the current build. This hides errors in other files in the playground
folder that might be out of date and have some errors we no longer
care about.

Hence the output is much more useful, not cluttered with irrelevant
error messages.

JIRA: RAIL-4273