Handle export result downloading better
In case the backend was constructed with a hostname, we must download
the export results from there. Since the results from the export
functions are relative URIs, we must prepend the hostname
explicitly. That way the downloads work properly.

This was not an issue previously, since KD is always hosted on the same
domain as the GoodData backend it uses for the data.

Also removed two useless hooks, they were not used and not exported.
There is a command for the same thing that should be used instead.

JIRA: RAIL-4387