Extract renderMode store slice
This is the first step to make the ui store slice really "UI-only":
without any business logic. Render mode violates this: there are
important business things triggered by changing of the render mode,
this it is not UI-only and must not be in the ui slice.

Also, the dependency cruiser rules were tightened to disallow touching
the selectors files directly from outside the model directory.

JIRA: RAIL-4170