RAIL-4460: Date filter calendar broken with SDK v8.10 plugin
Due to be able to use plugins with older version of SDK we need to keep css also for old definition of DatePicker. This is because plugins has not bundled css and uses parent css from dashboards.
This needs to be solved in future systematics way because can cause lots of problems ... :(

JIRA: RAIL-4460