Handle anonymous user on bear better
Anonymous user does not have the login MD5 so obtaining it would always
fail. Instead, we now fallback to empty permissions so that it does not
blow up anymore. This fixes live examples at the cost of the user no
longer being able to export the dashboard to PDF.

JIRA: RAIL-4596