Fix header Invite item white-labeling capabilities
The Invite item in application header had hard-coded font color. When headerColor property of the header is set to a custom color, Invite item did not respect it. The property is used in AD and KD for white-labeling feature. Additionally, icon of the item used SVG file directly and was not wrapped in React component that supports theming.

JIRA: TNT-1124