fix(sdk-ui-dashboard): prevent widget loads in view mode
Remove noop attribute filters in edit mode as they would cause a useless
query when a new filter (noop by default) is added. Keep them in view
mode so that switching to and from All on an ignored filter does not
show loading.

This is a tradeoff so that we optimize for the view mode performance
and also keep the more frequent use case in edit mode (adding a new
filter) loading-free as well. Switching to and from All in edit mode
will still show loading, but have no way of telling whether a noop
filter is noop because it was just added or because it was set that way
by the user.

JIRA: RAIL-4624