fix(sdk-ui-dashboard): fix dashboard cancel with plugin reload
Plugin reload can cause the insights collection to be missing some
insights that were on the original dashboard. So when cancelling, we
need to make sure that we have all the insights needed for the dashboard
state we are resetting to.

The resolution uses the caches so for 99% of the cases it should be
instantaneous.

JIRA: RAIL-4738