fix(sdk-ui-dashboard): make alert selector more tolerant
We must not blow up when a related widget is missing in the layout: it
can happen during edit mode interactions as we do not remove alerts from
the store. This makes cancelling the edit mode faster as the alerts do
not need to be loaded from the backend again.

Alternative would be to create a "persisted alerts" store in redux and
use it to back up the alerts. This is, however, so much work for such an
edge-case that I opted for this simpler solution instead.

JIRA: RAIL-4760