fix(sdk-ui-dashboard): save button behavior
In some cases, when you click the save button
(for example immediately after you insert an insight)
and the dashboard is still processing some requests
related to this insight, it can take few seconds,
before the CMD.SAVE starts to process
(as dashboard can process only 1 command at a time).
We want to show the loading indicator immediately and avoid any
double-save race conditions, so add a local state for this case.

JIRA: RAIL-4720