fix(sdk-ui-dashboard): fix formatting of drill warnings
Since the strings for the drill validation warnings use the translation
customization provider and gdc-dashboards has localization strings for
them with the same ID but different values, the version defined in
gdc-dashboards would win resulting in the insight names not being shown
in bold. To avoid this problem, we make the SDK message identifiers
unique in relation to the gdc-dashboards ones, thus using the correct
values for the formatting.

The reason for the translation localization provider being important
here is that it reuses some global state so that two instances of the
translation localization provider are not really independent and can
override each other messages if they have the same ids...

JIRA: RAIL-4716