fix: prevent null references in undoEnhancer
Since immer 9.0.13, the undoEnhancer breaks when the app is built with
react-scripts because the state parameter coming into the withUndo is
not a proper immer draft and thus the original() returns undefined.
In that case we now try to unwrap it from the proxy in an
immer-independent way to prevent null references down the road.

JIRA: RAIL-4751