fix: always expand Date dropdown list automatically
Open configuration widget first time always expand Date
dropdown list automatically

KD with new edit mode, In KD when opening configuration widget
of some insight for the first time, it always expands Date
dropdown list automatically. This is just a minor issue for
user but currently it is causing an issue for the automation
tests for toggling other attribute because this attribute is
covered by the Date dropdown list.

JIRA: RAIL-4817