feat(sdk-ui-dashboard): expose selectIsExport
Make this selector public as it can be required to implement
custom export logic in the dashboard plugins.

Also, all the sdk charts should be exported with isExportMode
in the chart configuration, or they can be rendered incorrectly
during the export.

Example: GeoPushpinChart in the dashboard plugin
without isExportMode in the config prop is not rendered at all.

JIRA: RAIL-4837