fix(dashboard-plugin-template): reselect version
Add override for reselect version in the plugin template,
because higher reselect version in the combination
with the current version of the TypeScript in the template
is causing typing errors for selectors,
that are created with createSelector function.

Also, update relevant logic in the plugin toolkit to change
overrides to resolutions, when using yarn package manager.

JIRA: RAIL-4837