feat: download exported report/dash via ObjectURL
Currently, exported report or dashboard task is polled via fetch API.
When task is not ready, JSON body with polling uri is returned.
When task finishes, export content is returned in response body.
Fetch utility always parsed the response body to text.

Now, when export task is polled, response body is parsed to
ArrayBuffer. When the export content is returned, buffer is converted
to Blob and then ObjectURL is created for the Blob. ObjectURL creation
attaches the blob to window instance and creates a URL that points to
it. This URL is then used in anchor element appended to DOM. Script
simulates the click on the anchor, file is downloaded to user's
filesystem, and anchor is removed from DOM, and ObjectURL revoked to
release blob data from window instance and memory.

This approach saves one additional call to backend when the content
of export is fetched again. Most importantly, it works around the
issues with authentication of this GET request made by click on the
anchor in the case when token authentication is used (no cookies) or
UserAgentReduction mode is enabled by the user's browser.

Name of the download must be parsed from the content-disposition
header, otherwise API uri value would be used.

JIRA: TNT-1344