feat: propagate created, updated, createdBy, updatedBy to tiger SPI
The following Tiger API entities got support for auditing properties:
- Analytical dashboards
- Analytical dashboards plugins
- Insights
- Metrics

The change propagates this information from entities to SPI. Unlike on
Bear, the map with cached users do not need to be managed as user
information is returned in "included" array of API response if request
asked for it to be included. The entity itself has "relationships" map
that points to "included" array.

Created/updated information is displayed in KD insight list now.

JIRA: TNT-1489