fix: reuse KPI execution for new alerts
Execution done just for Alert is exactly the same as for KPI itself,
only omits secondary metric.
It was probably meant as optimization,
but it caused cache miss and
one redundant execution for every KPI.
Even for ones which dont have alert at all.
For evaluation of new/updated alert KPI's execution is enough
as they share filters and secondary metric is ignored anyway.
JIRA: TNT-1716