fix: scatter plot bucket helper
Scatter plot: remove items from the viewBy and segmentBy bucket
that do not belong there (both can contain a maximum of 1 item).

This fix ensures that switching from different visualization
to scatter plot in AD will not cause the scatter plot
to contain invalid items in the viewBy and segmentBy bucket.

JIRA: F1-346