fix: heatmap coloring
Fix heatmap coloring issue.

Max value was out of the range specified for coloring,
because the applied number precision cut off the last few digits.

It seems that the number precision is not needed anymore,
Highcharts seems to handle the coloring correctly without it,
so remove it.

risk: low
JIRA: F1-344