feat: drill intersection filtering configuration
This features enables possibility to ignore specified attributes
in the drill intersection for selected drill definitions.
Ignored attributes are not transferred as filters during the drill.

Only drill to insight and drill to dashboards are supported.

- Add new feature flag `enableDrillIntersectionIgnoredAttributes`
to enable/disable drill intersection filtering configuration
- Add new components for drill intersection filtering configuration
- Apply new configuration in relevant command handlers
- Validate that ignored attributes are still present in target
insight, if not, remove the drill and show warning toast message
- Update some ui-kit components to support more customizations

risk: low
JIRA: F1-521