fix: better handling of fast changing filter prop
Remove check for status as it skipped fast changing filter prop.
In AD metric filter is taken from extended ref. point. So it is out of sync for short
time when filter is changed (waiting for PV loop).
This fast changing filter prop was ignored because status check.

JIRA: LX-615
risk: low