refactor: switch to secondary x axis for bar chart default sort
When the default Bar chart sorting is being generated,
the secondary_yaxis property is attempted to be retrieved
from the properties mapping.
Based on current behavior of UI, as well as the supportedProperties
definitions, it seems like the only secondary axis supported for
the bar chart is on the X axis, instead of the Y axis.

risk: low
JIRA: CQ-987