fix: widgets on the first row of container are not aligned
The first widget of the container is higher than the subsequent
widgets. We need to detect the row index of the widgets and gave
the widgets on the first row the same class and style them the same
way. The current code styled only the first widget of the first row
differently, counting with the fact that containers will have just
a single widget per row which was later changed.

JIRA: LX-736
risk: low