fix: performance
Add `enableCriticalContentPerformanceOptimizations` feature flag
to enable possible performance optimizations for the critical
content rendering of the dashboard.

Postpone initialization of the dashboard attribute filter(s),
which significantly reduces the time to show the visualizations,
especially in the case of a large number of attribute filters
(in case `enableCriticalContentPerformanceOptimizations` is enabled).

risk: low
JIRA: F1-1052