fix: propagate ad hoc displayAsLabel migrated attr filter immediately
If a filter was created before the introduction of the ‘support
duplicated label values’ feature, it is migrated to the new
configuration ad hoc during initialization. Previously, the new
configuration was only propagated to the filter’s parent application
when the user opened the filter, made changes, and applied them.

After the change, the migration is now detected during filter
initialization and reported immediately.

There are some consequences:
- If dashboard component starts in edit mode, Save button is enabled
  if filter was migrated. This does not happen in view mode.
- Filters can be stored in filter view in the old format when user
  save view quicker than all filters load values and mark themselves
  as "initialized: success"

JIRA: LX-824
risk: low