fix: do not enable Save button in edit mode for cross filters
The fix that allows saving of ad-hoc migrated filters in view mode
when user enters the edit mode enables Save button also when user
applied cross filter for an attribute with non-primary label. The fix
skips such filters to be considered as migrated.

This means that if filter is migrated in view mode and then
cross filter is applied on it, the filter will not be part of saved
set that can be saved in edit mode. However, this is not a problem
as user can migrate this filter the next time he will enter the edit
mode when the cross filter was not used.

JIRA: LX-846
risk: low