feat: support threshold measure in line chart
When ff is enabled, user can mark one of the line chart's measures
as a threshold one. The measure is not rendered in the chart but its
values are used to determine thresholds for other chart series.
The part of series for which threshold metrics returns 0, null, or
undefined are rendered in dashed line, the rest of the series is
rendered in solid line.

JIRA: LX-1072
risk: low