feat: change chart's X axis autorotation value from -90° to -45°
This change is required by MasterCard. It was decided to apply it
to every customer. Additionally, it is possible to select 45° rotation
manually from charts axis control in AD now.

Note that there seems to be a small visual bug when label values are
too long. When rotation value is set to Auto and rotation is handled
by the default HC's "autoRotation: -45" value, the long labels are
shortened correctly. When manual override is applied via
"rotation: -45", the right most label gets truncated by chart
viewport. This behavior seems to be caused by a bug in custom 3rd
party groupedCategories.ts plugin that we use and extend.

JIRA: LX-1057
risk: low